+++ modules/graph_filtering_module/graph_filtering.py
@@ -14,12 +14,13 @@ class GraphFiltering(QWidget):
         super().__init__()
 
         self.output_suffix = '_vGF'
+        self.celltracking_suffix = '_vTG'
 
         layout = QVBoxLayout()
 
         self.imagetypes = ['.nd2', '.tif', '.tiff', '.ome.tif', '.ome.tiff']
 
-        self.mask_graph_table = gf.FileTableWidget2(header_1="Mask", header_2="Graph", filenames_suffix_1='.ome.tif', filenames_suffix_2='.graphmlz')
+        self.mask_graph_table = gf.FileTableWidget2(header_1="Mask", header_2="Graph", filenames_suffix_1='.ome.tif', filenames_suffix_2='.graphmlz', filenames_filter=self.celltracking_suffix)
         self.mask_graph_table.file_table_changed.connect(self.mask_graph_table_changed)
         groupbox = QGroupBox('Segmentation masks and cell tracking graphs to process')
         layout2 = QVBoxLayout()